copied elf package from gopher2600
crunched package also copied. there is no snapshot system so this isn't
really needed but the SRAM type has been implemented with it in mind so
it's just as easy to use it

similar to the the copying of the arm package, references to the
environment package have been removed. instead the new context interface
is being used. not everything available in the old environment package
has been added to the context interface. there is no preferences and no
way of querying the TV type

constants from gopher2600's memorymap package have been replaced with
values local to the elf package

Snapshot() and Plumb() functions have been removed because there is no
rewind system in place yet

logging will always happen using the logging.Allow parameter

notification about unresolved symbols has been removed. in gopher2600
this is a method of informing the user via the GUI that something has
gone potentially wrong. there is no similar system yet in test7800 and
all the relevant information is in the log in any case