corrected GUI scaling to accomodate change in number of clocks per scanline
the change to the number of clocks per scanline was made in the previous
commit

TIA audio now ticks more slowly to accomodate the increased number of
clocks per scanline. added an IsAtari7800() function to the console
context to decide on how to count the audio sampling. this will be
worthwhile if we ever include 2600 emulation as an option alongside the
7800

comment correction in displaylist.go and removal of an unused constant
in elf.go