simplied arm T1 timer
(port from gopher2600)

elf now instruct ARM to clock() during immediate mode. for 2600 mappers CDF and
DPC+ the clock ticks differently when in immediate mode. that's not a concern
for 7800 emulation at the moment, but the changes have been included anyway

added ad-hoc "extended memory logging" option (in gopher2600 this is a proper
preference value)