refinement and explanation of preDMA value
added a delay to the writing of the background colour register. I'm
findin this necessary because the preDMA value has been determined to be
very precise and adjusting it further to accomodate the needs of
karateka (explained in the preDMA commentary) is impossible. this isn't
a krateka specific change, it affects every game

I'm not certain this is the precise mechanism by which the problem is
resolved but I believe that it must be something similar to this